// apps/angular-app/src/app/core/models/export-column-type.model.ts
export enum ExportColumnType {
  no = 'no',
  description = 'description',
  subtotal = 'subtotal',
  vat = 'vat',
  dibs = 'dibs',
  profitCosts = 'profitCosts',

  clsVat = 'clsVat',
  clsDibs = 'clsDibs',
  clsProfitCosts = 'clsProfitCosts',
  iPVat = 'iPVat',
  iPDibs = 'iPDibs',
  iPProfitCosts = 'iPProfitCosts',

  default = 'default',
  empty = 'empty',
}